test: update tests according to duplicated email error handling
